msvc: set warning level to /W4 globally
And fix a bunch of warnings

+++ src/dedicated_room/yuzu_room.cpp
@@ -368,9 +368,9 @@ int main(int argc, char** argv) {
     if (auto room = network.GetRoom().lock()) {
         AnnounceMultiplayerRoom::GameInfo preferred_game_info{.name = preferred_game,
                                                               .id = preferred_game_id};
-        if (!room->Create(room_name, room_description, bind_address, port, password, max_members,
-                          username, preferred_game_info, std::move(verify_backend), ban_list,
-                          enable_yuzu_mods)) {
+        if (!room->Create(room_name, room_description, bind_address, static_cast<u16>(port),
+                          password, max_members, username, preferred_game_info,
+                          std::move(verify_backend), ban_list, enable_yuzu_mods)) {
             LOG_INFO(Network, "Failed to create room: ");
             return -1;
         }
